requirements.txt
ebooklib>=0.18
markdown>=3.5
pyyaml>=6.0
Pillow>=10.0
